autolayout

+++ SportsCircleSportsCircle/SportsCircle/scheduleTableViewController.m
@@ -238,7 +238,7 @@
         NSString *objectCellId = object.objectId;
         //NSString *id1=[NSString stringWithFormat:@"%@",objectCellId];
         UITableViewCell * cell = [self.tableView cellForRowAtIndexPath:indexPath];
-        NSLog(@"cell title:%@ , row:%li and objectId:%@",cell.textLabel.text,indexPath.row,objectCellId);
+        NSLog(@"cell title:%@ , row:%li and objectId:%@",cell.textLabel.text,(long)indexPath.row,objectCellId);
         
         //NSDate *object = datas[indexPath.row];
         scheduleEditViewController *controller = (scheduleEditViewController *)[segue destinationViewController];
